# background/docs/openapi3.json
{
  "components": {
    "schemas": {
      "types.Column": {
        "properties": {
          "exp": {
            "description": "expressions, which default to = when the value is null, have =, !=, \u003e, \u003e=, \u003c, \u003c=, like",
            "type": "string"
          },
          "logic": {
            "description": "logical type, default value is \"and\", support \u0026, and, ||, or",
            "type": "string"
          },
          "name": {
            "description": "column name",
            "type": "string"
          },
          "value": {
            "description": "column value"
          }
        },
        "type": "object"
      },
...
      "types.CreateMenusReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "properties": {
              "id": {
                "description": "id",
                "type": "integer"
              }
            },
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
      "types.CreateMenusRequest": {
        "properties": {
          "icon": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "order": {
            "type": "integer"
          },
          "parentID": {
            "type": "integer"
          },
          "path": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.DeleteMenusByIDReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.GetMenusByIDReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "properties": {
              "menus": {
                "$ref": "#/components/schemas/types.MenusObjDetail"
              }
            },
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.ListMenussReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "properties": {
              "menuss": {
                "items": {
                  "$ref": "#/components/schemas/types.MenusObjDetail"
                },
                "type": "array"
              }
            },
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.MenusObjDetail": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "id": {
            "description": "convert to uint64 id",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "order": {
            "type": "integer"
          },
          "parentID": {
            "type": "integer"
          },
          "path": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "types.Params": {
        "properties": {
          "columns": {
            "description": "query conditions",
            "items": {
              "$ref": "#/components/schemas/types.Column"
            },
            "type": "array"
          },
          "limit": {
            "description": "lines per page",
            "type": "integer"
          },
          "page": {
            "description": "page number, starting from page 0",
            "type": "integer"
          },
          "sort": {
            "description": "sorted fields, multi-column sorting separated by commas",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "types.UpdateMenusByIDReply": {
        "properties": {
          "code": {
            "description": "return code",
            "type": "integer"
          },
          "data": {
            "description": "return data",
            "type": "object"
          },
          "msg": {
            "description": "return information description",
            "type": "string"
          }
        },
        "type": "object"
      },
      "types.UpdateMenusByIDRequest": {
        "properties": {
          "icon": {
            "type": "string"
          },
          "id": {
            "description": "uint64 id",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "order": {
            "type": "integer"
          },
          "parentID": {
            "type": "integer"
          },
          "path": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Type Bearer your-jwt-token to Value",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "http server api docs",
    "title": "godemo api docs",
    "version": "v1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/menus": {
      "post": {
        "description": "Creates a new menus entity using the provided data in the request body.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/types.CreateMenusRequest"
              }
            }
          },
          "description": "menus information",
          "required": true,
          "x-originalParamName": "data"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.CreateMenusReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a new menus",
        "tags": [
          "menus"
        ]
      }
    },
    "/api/v1/menus/list": {
      "post": {
        "description": "Returns a paginated list of menus based on query filters, including page number and size.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/types.Params"
              }
            }
          },
          "description": "query parameters",
          "required": true,
          "x-originalParamName": "data"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.ListMenussReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get a paginated list of menuss by custom conditions",
        "tags": [
          "menus"
        ]
      }
    },
    "/api/v1/menus/{id}": {
      "delete": {
        "description": "Deletes a existing menus identified by the given id in the path.",
        "parameters": [
          {
            "description": "id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.DeleteMenusByIDReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete a menus by id",
        "tags": [
          "menus"
        ]
      },
      "get": {
        "description": "Gets detailed information of a menus specified by the given id in the path.",
        "parameters": [
          {
            "description": "id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.GetMenusByIDReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get a menus by id",
        "tags": [
          "menus"
        ]
      },
      "put": {
        "description": "Updates the specified menus by given id in the path, support partial update.",
        "parameters": [
          {
            "description": "id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/types.UpdateMenusByIDRequest"
              }
            }
          },
          "description": "menus information",
          "required": true,
          "x-originalParamName": "data"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/types.UpdateMenusByIDReply"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update a menus by id",
        "tags": [
          "menus"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    },
    {
      "url": "https://localhost:8080/"
    }
  ]
}